add base token grain
