+++ ManagedCode.Orleans.Identity.Server/Constants/TokenGrainConstants.cs
@@ -4,5 +4,6 @@
     {
         public const string EMAIL_VERIFICATION_TOKEN_REMINDER_NAME = "DisableEmailVerificationToken";
         public const string MAGIC_LINK_TOKEN_REMINDER_NAME = "DisableMagicLinkToken";
+        public const string CODE_VERIFICATION_TOKEN_REMINDER_NAME = "verificationCodeToken";
     }
 }
